+++ lib/aliyun/log/record/exception.rb
@@ -4,6 +4,7 @@ module Aliyun
       class ArgumentError < StandardError; end
       class UnknownAttributeError < StandardError; end
       class ProjectNameError < StandardError; end
+      class ParseStatementInvalid < StandardError; end
     end
   end
 end